move test cases to cached_store_test
