--- pkg/chunk/store_test.go
@@ -1,239 +1,0 @@
-/*
- * JuiceFS, Copyright (C) 2020 Juicedata, Inc.
- *
- * This program is free software: you can use, redistribute, and/or modify
- * it under the terms of the GNU Affero General Public License, version 3
- * or later ("AGPL"), as published by the Free Software Foundation.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses/>.
- */
-
-//nolint:errcheck
-package chunk
-
-import (
-	"bytes"
-	"context"
-	"io"
-	"os"
-	"path/filepath"
-	"testing"
-	"time"
-
-	"github.com/juicedata/juicefs/pkg/object"
-)
-
-func forgeChunk(store ChunkStore, chunkid uint64, size int) error {
-	w := store.NewWriter(chunkid)
-	buf := bytes.Repeat([]byte{0x41}, size)
-	if _, err := w.WriteAt(buf, 0); err != nil {
-		return err
-	}
-	return w.Finish(size)
-}
-
-func testStore(t *testing.T, store ChunkStore) {
-	writer := store.NewWriter(1)
-	data := []byte("hello world")
-	if n, err := writer.WriteAt(data, 0); n != 11 || err != nil {
-		t.Fatalf("write fail: %d %s", n, err)
-	}
-	offset := defaultConf.BlockSize - 3
-	if n, err := writer.WriteAt(data, int64(offset)); err != nil || n != 11 {
-		t.Fatalf("write fail: %d %s", n, err)
-	}
-	if err := writer.FlushTo(defaultConf.BlockSize + 3); err != nil {
-		t.Fatalf("flush fail: %s", err)
-	}
-	size := offset + len(data)
-	if err := writer.Finish(size); err != nil {
-		t.Fatalf("finish fail: %s", err)
-	}
-	defer store.Remove(1, size)
-
-	reader := store.NewReader(1, size)
-	p := NewPage(make([]byte, 5))
-	if n, err := reader.ReadAt(context.Background(), p, 6); n != 5 || err != nil {
-		t.Fatalf("read failed: %d %s", n, err)
-	} else if string(p.Data[:n]) != "world" {
-		t.Fatalf("not expected: %s", string(p.Data[:n]))
-	}
-	p = NewPage(make([]byte, 20))
-	if n, err := reader.ReadAt(context.Background(), p, offset); n != 11 || err != nil && err != io.EOF {
-		t.Fatalf("read failed: %d %s", n, err)
-	} else if string(p.Data[:n]) != "hello world" {
-		t.Fatalf("not expected: %s", string(p.Data[:n]))
-	}
-
-	bsize := defaultConf.BlockSize / 2
-	errs := make(chan error, 3)
-	for i := 2; i < 5; i++ {
-		go func(chunkid uint64) {
-			if err := forgeChunk(store, chunkid, bsize); err != nil {
-				errs <- err
-				return
-			}
-			time.Sleep(time.Millisecond * 100) // waiting for flush
-			errs <- store.Remove(chunkid, bsize)
-		}(uint64(i))
-	}
-	for i := 0; i < 3; i++ {
-		if err := <-errs; err != nil {
-			t.Fatalf("test concurrent write failed: %s", err)
-		}
-	}
-}
-
-var defaultConf = Config{
-	BlockSize:  1 << 20,
-	CacheDir:   filepath.Join(os.TempDir(), "diskCache"),
-	CacheSize:  1,
-	MaxUpload:  1,
-	PutTimeout: time.Second,
-	GetTimeout: time.Second * 2,
-	AutoCreate: true,
-	BufferSize: 10 << 20,
-}
-
-func TestStoreDefault(t *testing.T) {
-	mem, _ := object.CreateStorage("mem", "", "", "")
-	_ = os.RemoveAll(defaultConf.CacheDir)
-	store := NewCachedStore(mem, defaultConf)
-	testStore(t, store)
-	if used := store.UsedMemory(); used != 0 {
-		t.Fatalf("used memory %d != expect 0", used)
-	}
-	if cnt, used := store.(*cachedStore).bcache.stats(); cnt != 0 || used != 0 {
-		t.Fatalf("cache cnt %d used %d, expect both 0", cnt, used)
-	}
-}
-
-func TestStoreMemCache(t *testing.T) {
-	mem, _ := object.CreateStorage("mem", "", "", "")
-	conf := defaultConf
-	conf.CacheDir = "memory"
-	store := NewCachedStore(mem, conf)
-	testStore(t, store)
-	if used := store.UsedMemory(); used != 0 {
-		t.Fatalf("used memory %d != expect 0", used)
-	}
-	if cnt, used := store.(*cachedStore).bcache.stats(); cnt != 0 || used != 0 {
-		t.Fatalf("cache cnt %d used %d, expect both 0", cnt, used)
-	}
-}
-func TestStoreCompressed(t *testing.T) {
-	mem, _ := object.CreateStorage("mem", "", "", "")
-	conf := defaultConf
-	conf.Compress = "lz4"
-	conf.AutoCreate = false
-	store := NewCachedStore(mem, conf)
-	testStore(t, store)
-}
-
-func TestStoreLimited(t *testing.T) {
-	mem, _ := object.CreateStorage("mem", "", "", "")
-	conf := defaultConf
-	conf.UploadLimit = 1 << 20
-	conf.DownloadLimit = 1 << 20
-	store := NewCachedStore(mem, conf)
-	testStore(t, store)
-}
-
-func TestStoreFull(t *testing.T) {
-	mem, _ := object.CreateStorage("mem", "", "", "")
-	conf := defaultConf
-	conf.FreeSpace = 0.9999
-	store := NewCachedStore(mem, conf)
-	testStore(t, store)
-}
-
-func TestStoreSmallBuffer(t *testing.T) {
-	mem, _ := object.CreateStorage("mem", "", "", "")
-	conf := defaultConf
-	conf.BufferSize = 1 << 20
-	store := NewCachedStore(mem, conf)
-	testStore(t, store)
-}
-
-func TestStoreAsync(t *testing.T) {
-	mem, _ := object.CreateStorage("mem", "", "", "")
-	conf := defaultConf
-	conf.Writeback = true
-	p := filepath.Join(conf.CacheDir, stagingDir, "chunks/0/0/123_0_4")
-	os.MkdirAll(filepath.Dir(p), 0744)
-	f, _ := os.Create(p)
-	f.WriteString("good")
-	f.Close()
-	store := NewCachedStore(mem, conf)
-	time.Sleep(time.Millisecond * 50) // wait for scan to finish
-	in, err := mem.Get("chunks/0/0/123_0_4", 0, -1)
-	if err != nil {
-		t.Fatalf("staging object should be upload")
-	}
-	data, _ := io.ReadAll(in)
-	if string(data) != "good" {
-		t.Fatalf("data %s != expect good", data)
-	}
-	testStore(t, store)
-}
-
-func TestStoreDelayed(t *testing.T) {
-	mem, _ := object.CreateStorage("mem", "", "", "")
-	conf := defaultConf
-	conf.Writeback = true
-	conf.UploadDelay = time.Millisecond * 200
-	store := NewCachedStore(mem, conf)
-	time.Sleep(time.Second) // waiting for cache scanned
-	testStore(t, store)
-	if err := forgeChunk(store, 10, 1024); err != nil {
-		t.Fatalf("forge chunk 10 1024: %s", err)
-	}
-	defer store.Remove(10, 1024)
-	time.Sleep(time.Second) // waiting for upload
-	if _, err := mem.Head("chunks/0/0/10_0_1024"); err != nil {
-		t.Fatalf("head object 10_0_1024: %s", err)
-	}
-}
-
-func TestStoreMultiBuckets(t *testing.T) {
-	mem, _ := object.CreateStorage("mem", "", "", "")
-	conf := defaultConf
-	conf.Partitions = 3
-	store := NewCachedStore(mem, conf)
-	testStore(t, store)
-}
-
-func TestFillCache(t *testing.T) {
-	mem, _ := object.CreateStorage("mem", "", "", "")
-	_ = os.RemoveAll(defaultConf.CacheDir)
-	store := NewCachedStore(mem, defaultConf)
-	bsize := defaultConf.BlockSize
-	if err := forgeChunk(store, 10, 1024); err != nil {
-		t.Fatalf("forge chunk 10 1024: %s", err)
-	}
-	defer store.Remove(10, 1024)
-	if err := forgeChunk(store, 11, bsize); err != nil {
-		t.Fatalf("forge chunk 11 %d: %s", bsize, err)
-	}
-	defer store.Remove(11, bsize)
-
-	bcache := store.(*cachedStore).bcache
-	if cnt, used := bcache.stats(); cnt != 1 || used != 1024+4096 { // only chunk 10 cached
-		t.Fatalf("cache cnt %d used %d, expect cnt 1 used 5120", cnt, used)
-	}
-	if err := store.FillCache(10, 1024); err != nil {
-		t.Fatalf("fill cache 10 1024: %s", err)
-	}
-	if err := store.FillCache(11, uint32(bsize)); err != nil {
-		t.Fatalf("fill cache 11 %d: %s", bsize, err)
-	}
-	expect := int64(1024 + 4096 + bsize)
-	if cnt, used := bcache.stats(); cnt != 2 || used != expect {
-		t.Fatalf("cache cnt %d used %d, expect cnt 2 used %d", cnt, used, expect)
-	}
-}
